Move contents of extra_colors to new module color_consts
